Update FiveM install script
This moves to pulling the latest version for FiveM again.

If they enable cloudflare protection it will break but it has support for download URL.

DOWNLOAD_URL is still limited to fx.tar.xz being the filename.

The installer will fail if the file is not named correctly or is the wrong file type.