Allow EXILED Modification to run Auto Updater
I'm proposing this change as it will add an Environment Variable used by the EXILED (a Popular SCP:SL Mods Framework) to run its auto updater. Otherwise, the user could run into an issue with his server not starting.

I know this is normally not used to support modifications. But as this is only an environment variable I will propose the change anyway.