Updating Minetest Egg for Pterodactyl's Productivity Improvements and Default Versioning
Hiya there! I was tired of having a "half ash" setup that this "stock" egg supplied so I "messed" with the .json and came up with the following... 
Added --terminal so the server ACTUALLY starts in a working terminal just like Minecraft servers do (you can either talk in it without the "/" or execute commands as the set admin).
Changed Stop command from the "^C" now that "/shutdown" can be used because of a working terminal
Changed "done" to "Server for gameid" because it wasn't working after these changes for " listening on " on my end. 
Updated Versioning to 5.4.1 instead of now dated 5.3.0

Caveat: Slightly annoying but the game requires you to input a name for the Admin User by the use of putting "name=youradmin" into the minetest.conf of the instance. As well as logging in as that admin WITH a ingame password. But hey that's progress I am not right?

Hope my hours helps those who were looking for a Minetest egg that is more productive as well up to date!