fix: sogebot install for sogebot v13+
Sogebot now requires Nodejs 16 to build and run and is not backward compatible with previous versions. Didn't add backwards compatibility because of this, but we might have to add version detection for building similar to Spigot?

Migrated MATCH and PACKAGE variables into the install script itself to make it cleaner. They're not something the end user needs to see.

resolves #1271